2022-01-16_우형래 :[modify] findmate (수정 index 페이지, 상세페이지 반응형 UI) , rundaycourse (반응형 UI)
