+++ src/components/units/FindMates/add/FindMatesAddQueries.ts
@@ -24,3 +24,23 @@ export const UPDATE_BOARD = gql`
     }
   }
 `;
+
+export const FETCH_BOARD = gql`
+  query fetchBoard($boardId: ID!) {
+    fetchBoard(boardId: $boardId) {
+      _id
+      writer
+      title
+      contents
+      youtubeUrl
+      likeCount
+      images
+      boardAddress {
+        zipcode
+        address
+        addressDetail
+      }
+      createdAt
+    }
+  }
+`;
